# docs/api_spec_openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Parks Connect API",
    "version": "1.0.0",
    "description": "REST API for Parks Connect covering auth, visitor logs, environmental logs, feedback, notifications, and analytics."
  },
  "servers": [
    { "url": "http://localhost:4000", "description": "Local dev" }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/feedback/{id}/status": {
      "put": {
        "summary": "Update feedback status",
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "properties": { "status": { "type": "string" } } } } }
        },
        "responses": { "200": { "description": "Updated" } }
      }
    },
...
  }
}